backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.44.0
python-dotenv==1.1.0
pydantic==2.13.2
openai==2.32.0
langchain-core==1.3.0
langchain-chroma==1.1.0
langchain-huggingface==1.2.2
langchain-text-splitters==1.1.2
chromadb==1.5.8
pypdf
loguru